src/test/kotlin/com/yonatankarp/adventofcode/aoc2022/day10/Day10Test.kt
package com.yonatankarp.adventofcode.aoc2022.day10

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test
import com.yonatankarp.adventofcode.utils.resourceAsList
import org.junit.jupiter.api.Nested

class Day10Test {

    private val exampleInput = """
        addx 15
        addx -11
        addx 6
        addx -3
        addx 5
        addx -1
        addx -8
        addx 13
        addx 4
        noop
        addx -1
        addx 5
        addx -1
        addx 5
        addx -1
        addx 5
        addx -1
        addx 5
        addx -1
        addx -35
        addx 1
        addx 24
        addx -19
        addx 1
        addx 16
        addx -11
        noop
        noop
        addx 21
        addx -15
        noop
        noop
        addx -3
        addx 9
        addx 1
        addx -3
        addx 8
        addx 1
        addx 5
        noop
        noop
        noop
        noop
        noop
        addx -36
        noop
        addx 1
        addx 7
        noop
        noop
        noop
        addx 2
        addx 6
        noop
        noop
        noop
        noop
        noop
        addx 1
        noop
        noop
        addx 7
        addx 1
        noop
        addx -13
        addx 13
        addx 7
        noop
        addx 1
        addx -33
        noop
        noop
        noop
        addx 2
        noop
        noop
        noop
        addx 8
        noop
        addx -1
        addx 2
        addx 1
        noop
        addx 17
        addx -9
        addx 1
        addx 1
        addx -3
        addx 11
        noop
        noop
        addx 1
        noop
        addx 1
        noop
        noop
        addx -13
        addx -19
        addx 1
        addx 3
        addx 26
        addx -30
        addx 12
        addx -1
        addx 3
        addx 1
        noop
        noop
        noop
        addx -9
        addx 18
        addx 1
        addx 2
        noop
        noop
        addx 9
        noop
        noop
        noop
        addx -1
        addx 2
        addx -37
        addx 1
        addx 3
        noop
        addx 15
        addx -21
        addx 22
        addx -6
        addx 1
        noop
        addx 2
        addx 1
        noop
        addx -10
        noop
        noop
        addx 20
        addx 1
        addx 2
        addx 2
        addx -6
        addx -11
        noop
        noop
        noop
    """.trimIndent().split("\n")

    @Nested
    inner class Part1 {
        @Test
        fun `Part 1 - Example`() {
            val day10 = Day10(exampleInput)
            assertEquals(13140, day10.solvePart1())
        }

        @Test
        fun `Part 1 - Real Input`() {
            val day10 = Day10(resourceAsList("2022/day10.txt"))
            assertEquals(15260, day10.solvePart1())
        }
    }

    @Nested
    inner class Part2 {
        @Test
        fun `Part 2 - Example`() {
            val exampleOutput = """
                ##..##..##..##..##..##..##..##..##..##..
                ###...###...###...###...###...###...###.
                ####....####....####....####....####....
                #####.....#####.....#####.....#####.....
                ######......######......######......####
                #######.......#######.......#######.....
            """.trimIndent()
            val day10 = Day10(exampleInput)
            assertEquals(exampleOutput, day10.solvePart2())
        }

        @Test
        fun `Part 2 - Real Input`() {
            val realOutput = """
                ###...##..#..#.####..##..#....#..#..##..
                #..#.#..#.#..#.#....#..#.#....#..#.#..#.
                #..#.#....####.###..#....#....#..#.#....
                ###..#.##.#..#.#....#.##.#....#..#.#.##.
                #....#..#.#..#.#....#..#.#....#..#.#..#.
                #.....###.#..#.#.....###.####..##...###.
            """.trimIndent()
            val day10 = Day10(resourceAsList("2022/day10.txt"))
            assertEquals(realOutput, day10.solvePart2())
        }
    }
}
